rt: reduce code defined in macros
Instead of defining code in macros, move code definition to sub modules
and use the cfg_macro to declare the module.
