+++ tokio/src/runtime/scheduler/multi_thread/handle.rs
@@ -9,6 +9,14 @@ use crate::util::RngSeedGenerator;
 
 use std::fmt;
 
+cfg_metrics! {
+    mod metrics;
+}
+
+cfg_taskdump! {
+    mod taskdump;
+}
+
 /// Handle to the multi thread scheduler
 pub(crate) struct Handle {
     /// Task spawner
@@ -50,73 +58,6 @@ impl Handle {
         }
 
         handle
-    }
-}
-
-cfg_metrics! {
-    use crate::runtime::{SchedulerMetrics, WorkerMetrics};
-
-    impl Handle {
-        pub(crate) fn num_workers(&self) -> usize {
-            self.shared.worker_metrics.len()
-        }
-
-        pub(crate) fn num_blocking_threads(&self) -> usize {
-            self.blocking_spawner.num_threads()
-        }
-
-        pub(crate) fn num_idle_blocking_threads(&self) -> usize {
-            self.blocking_spawner.num_idle_threads()
-        }
-
-        pub(crate) fn active_tasks_count(&self) -> usize {
-            self.shared.owned.active_tasks_count()
-        }
-
-        pub(crate) fn scheduler_metrics(&self) -> &SchedulerMetrics {
-            &self.shared.scheduler_metrics
-        }
-
-        pub(crate) fn worker_metrics(&self, worker: usize) -> &WorkerMetrics {
-            &self.shared.worker_metrics[worker]
-        }
-
-        pub(crate) fn injection_queue_depth(&self) -> usize {
-            self.shared.injection_queue_depth()
-        }
-
-        pub(crate) fn worker_local_queue_depth(&self, worker: usize) -> usize {
-            self.shared.worker_local_queue_depth(worker)
-        }
-
-        pub(crate) fn blocking_queue_depth(&self) -> usize {
-            self.blocking_spawner.queue_depth()
-        }
-    }
-}
-
-cfg_taskdump! {
-    impl Handle {
-        pub(crate) async fn dump(&self) -> crate::runtime::Dump {
-            let trace_status = &self.shared.trace_status;
-
-            // If a dump is in progress, block.
-            trace_status.start_trace_request(&self).await;
-
-            let result = loop {
-                if let Some(result) = trace_status.take_result() {
-                    break result;
-                } else {
-                    self.notify_all();
-                    trace_status.result_ready.notified().await;
-                }
-            };
-
-            // Allow other queued dumps to proceed.
-            trace_status.end_trace_request(&self).await;
-
-            result
-        }
     }
 }
