+++ tokio/src/runtime/scheduler/multi_thread/worker/taskdump_mock.rs
@@ -1,0 +1,7 @@
+use super::{Core, Handle};
+
+impl Handle {
+    pub(super) fn trace_core(&self, core: Box<Core>) -> Box<Core> {
+        core
+    }
+}
